pyproject.toml
[tool.poetry]
name = "receiptwallet"
version = "0.1.0"
description = ""
authors = ["Johannes <joh@nnes.fi>"]

[tool.poetry.dependencies]
python = "^3.10"
pytesseract = "^0.3.10"
opencv-python = "^4.8.0"
numpy = "^1.25.2"
Pillow = "^10.0.0"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
pydantic = "^2.3.0"
python-multipart = "^0.0.6"
python-jose = "3.3.0"
passlib = "^1.7.4"
SQLAlchemy = "^2.0.20"
PyYAML = "^6.0.1"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
